Shift logic to post synthetic class step, up into make_synthetic_system_network
